Hazard detection works!!
Hazard detection works, run the program in the text file to test it.
Branching has been implemented BUT HAS NOT BEEN TESTED.
Even though hazard detection (with NOPs) works, it does not follow the
correct data path. It stalls the processes in their buffers instead of
having a NOP pass through the pipeline. We will need to change this at
some point, but at least it works for now.
Remember to use the python scripts, they help a ton!